Move graphql filter utils to separate file
