+++ saleor/graphql/product/resolvers.py
@@ -2,7 +2,8 @@ from django.db.models import Sum
 
 from ...order import OrderStatus
 from ...product import models
-from ..utils import filter_by_period, get_database_id, get_user_or_app_from_context
+from ..utils import get_database_id, get_user_or_app_from_context
+from ..utils.filters import filter_by_period
 from .filters import (
     filter_attributes_by_product_types,
     filter_products_by_stock_availability,
