+++ saleor/graphql/webhook/filters.py
@@ -2,7 +2,7 @@ import django_filters
 
 from ...webhook.models import Webhook
 from ..core.types import FilterInputObjectType
-from ..utils import filter_by_query_param
+from ..utils.filters import filter_by_query_param
 
 
 def filter_webhook_search(qs, _, value):
